Move annotations to their own folder
