+++ ui/src/features/annotator/annotations/annotation-shape.tsx
@@ -1,8 +1,8 @@
 // Copyright (C) 2025 Intel Corporation
 // SPDX-License-Identifier: Apache-2.0
 
-import { useAnnotator } from './annotator-provider.component';
-import { Annotation, Point } from './types';
+import { useAnnotator } from '../annotator-provider.component';
+import { Annotation, Point } from '../types';
 
 type AnnotationShapeProps = {
     annotation: Annotation;
