+++ ui/src/features/annotator/annotations/annotation.component.tsx
@@ -1,8 +1,8 @@
 // Copyright (C) 2025 Intel Corporation
 // SPDX-License-Identifier: Apache-2.0
 
+import { Annotation as AnnotationInterface } from '../types';
 import { AnnotationShape } from './annotation-shape';
-import { Annotation as AnnotationInterface } from './types';
 
 interface AnnotationProps {
     annotation: AnnotationInterface;
