+++ ui/src/features/annotator/annotations/mask-annotations.tsx
