+++ ui/src/features/annotator/annotator-canvas.tsx
@@ -8,7 +8,7 @@ import thumbnailUrl from '../../assets/mocked-project-thumbnail.png';
 import { ZoomProvider } from '../../components/zoom/zoom';
 import { ZoomTransform } from '../../components/zoom/zoom-transform';
 import { response } from '../dataset/mock-response';
-import { Annotations } from './annotations';
+import { Annotations } from './annotations/annotations';
 import { ToolManager } from './tools/tool-manager.component';
 
 type Item = (typeof response.items)[number];
